requirements.txt
pyyaml
elasticsearch
Flask
itsdangerous
Jinja2
MarkupSafe
requests
urllib3
virtualenv
Werkzeug
